Introduce a local-first Environment Designer feature with contracts, application services, runtime projection, publish/rollback, and a WPF admin surface. Adds EnvironmentDesign DTOs, a full EnvironmentDesignerService implementation and interfaces, new UI page/views/viewmodels, endpoint routes under /api/environment-designer, and DI registration for the service. Updates storage registration, provider config loading to use prepared config, and routes CLI config validation through the configuration pipeline. Also improves the desktop launcher to stop workspace Meridian.Desktop.exe instances before building, support a fixture desktop launcher mode, and updates docs and tests to cover the new environment designer and related UI/navigation/catalog changes.

+++ src/Meridian.Execution/Models/OrderStatusUpdate.cs
@@ -9,7 +9,7 @@ public sealed record OrderStatusUpdate(
     string ClientOrderId,
     string Symbol,
     OrderStatus Status,
-    long FilledQuantity,
+    decimal FilledQuantity,
     decimal? AverageFillPrice,
     string? RejectReason,
     DateTimeOffset Timestamp);
